-UI: -StartRobot button added      -StopMotor Button added
-mainPageViewModel:
     -Scan method added
     -StartRobot method added TODO
     -StopRobot method added
     -MeasureFilter fixed

     -public boolean stop added
     -Scan2 minor changes

+++ MeasureController/MeasureController/Models/Measure.cs
@@ -4,12 +4,12 @@
     {
         public double RobotTheta { get; set; }
 
-        public double SensorMotorPosition { get; set; }
+        public float SensorMotorPosition { get; set; }
 
         public double RobotX { get; set; }
 
         public double RobotY { get; set; }
 
-        public double SensorDistance { get; set; }
+        public float SensorDistance { get; set; }
     }
 }
